Refactored and renamed files
