+++ packages/retail-ui-extensions/src/components/Types/auto-capitalization-type.ts
