+++ packages/retail-ui-extensions/src/components/index.ts
@@ -8,4 +8,4 @@ export * from './SegmentedControl';
 export * from './Tag';
 export * from './Dialog';
 export * from './SearchBar';
-export * from './Types';
+export * from './types';
