--- packages/retail-ui-extensions/src/extension-api/CartApi/index.ts
@@ -1,12 +1,0 @@
-import type {CartApiContent} from './CartApiContent';
-
-/**
- * Access the merchant’s current cart.
- */
-export interface CartApi {
-  cart: CartApiContent;
-}
-
-export function isCartApi(api: any): api is CartApi {
-  return 'cart' in api;
-}
