--- packages/retail-ui-extensions/src/extension-api/Models/index.ts
@@ -1,1 +1,0 @@
-export {Cart, Customer, Discount, LineItem, CustomSale} from './Cart';
