+++ packages/retail-ui-extensions/src/extension-api/cart-api/index.ts
@@ -1,0 +1,8 @@
+import type {CartApiContent} from './cart-api-content';
+
+/**
+ * Access the merchant’s current cart.
+ */
+export interface CartApi {
+  cart: CartApiContent;
+}
