+++ packages/retail-ui-extensions/src/extension-api/locale-api/index.ts
@@ -1,0 +1,8 @@
+import {LocaleApiContent} from './locale-api-content';
+
+/**
+ * Access the merchant’s current locale (in [IETF format](https://en.wikipedia.org/wiki/IETF_language_tag)) to internationalize your extension content.
+ */
+export interface LocaleApi {
+  locale: LocaleApiContent;
+}
