+++ packages/retail-ui-extensions/src/extension-api/locale-api/locale-api-content.ts
@@ -1,0 +1,7 @@
+export interface LocaleApiContent {
+  /** IETF-formatted locale at time of page load */
+  initialValue: string;
+
+  /** Callback when user changes locale during runtime */
+  setOnChange(onChange: (locale: string) => void): void;
+}
