+++ packages/retail-ui-extensions/src/extension-api/navigation-api/index.ts
@@ -1,0 +1,1 @@
+export type {NavigationApiContent, NavigationApi} from './navigation-api';
