+++ packages/retail-ui-extensions/src/extension-api/session-token-api/index.ts
@@ -1,0 +1,1 @@
+export type {SessionTokenApiContent, SessionTokenApi} from './session-api';
