+++ packages/retail-ui-extensions/src/extension-api/types/index.ts
@@ -1,0 +1,1 @@
+export {Address, Cart, Customer, Discount, LineItem, CustomSale} from './cart';
